pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "alloy"
version = "0.1.0"
description = "Hybrid SSM-Attention language model on Apple Silicon with MLX"
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    "mlx>=0.22.0",
    "numpy>=1.24",
    "pyyaml>=6.0",
    "tiktoken>=0.5",
    "sentencepiece>=0.1.99",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[tool.setuptools.packages.find]
include = ["alloy*"]
